#163118
Add identifier to BibleRefs to indicate that the Scripture reference is
primary to the article; referenced at the beginning of the article, i.e.
was read at the beginning of the meeting, and is therefore central to
the article.  This is to enable searching by 'primary' scriptures.
Also, add the Article Page Number to the BibleRefs to enable linking of
articles to all their corresponding BibleRefs.